Boom Boom now has an attack functionality
Single press of the left mouse buttons results in a simple attack, holding it will result in a strong attack. The animations executed are sample ones, they are about to be updated with real ones. Boom Boom class also has been updated to pass the "UPlayerInput" component of its player character to Zip Zap so he can process the input as well and use it to update his states and animations.