Trash now *actually* does damage, toxic damage has been adjusted, characters heal after a while when idle
Forgot to actually make the trash do damage which has been fixed, and the damage of the toxic obstacle has been lowered. The hurt sprites for Boom Boom and Zip Zap have also been added for future use. The characters will also now heal after standing still for about 10 seconds, but maybe this should be decreased or increased in the future.

- Joe