Added functionality for two controllers
Boom Boom is the first spawner character, zip zap is the second. The first PLAYER controller (not physical) reads the input from the mouse, keyboard and the first ever connected additional USB controller (all 3 are physical). The second connected USB controller (a device) automatically gets assigned to the second PLAYER controller (not physical).