Migrated the level blueprint to C++, fixed bugs and redesigned system for level switching
The functionality in the level blueprint is now entirely migrated in a C++ class called "LevelManager". The functionality of switching between levels is completely redesigned and ability to move the camera and the character between two levels isn't limited only to the X axis, but now can move along the Z one as well. The positions for each level for both the camera and the characters' spawn locations (usually it's just one location for both, zip zap is just spawned 120 (don't know which unit to use... let's say cm))next to boom boom) are initialized in the "InitializeCameraLocations" and "InitializeLevelStartLocations" functions, respectively.

Bugs in the behavior of the mouse and the particles' spawning are fixed, but locations for these spawnings still need to be touched a little bit.